school project
we all had to make an escaperoom we were allowed to make it in English, but most of my classmates did their escape room in Dutch and i worked very hard for this one because it was about freeing my country, at the end i got the highest score for my escape room out of my class a 9,7 out of 10. 
Last this please do not share this to a lot of people because me and my family could get in trouble because it is against Turkish law.